refactor: UserResponse 타입 재설정
